Sensey 15.21 - 15.30 qo tramadrutyan tak e

+++ java-core-lessons/src/com/demo/progresstests/exercies15/Exercies15_21.java
@@ -12,7 +12,7 @@ public class Exercies15_21 {
         Scanner scanner = new Scanner(System.in);
 
         System.out.println("number");
-        long num = scanner.nextLong();
+        int  num = scanner.nextInt();
 
         while (true) {
             int sum = 0;
